refactor: seperate newItemCreation from explorerFolder
